Clean up missing or extra properties in GDAXFeed.
Switch GDAXFeed#mapMessage() and GDAXFeed#mapAuthMessage() from using
type assertions to declared types as otherwise the compiler will not
complain about missing or extra properties.

For UnknownMessage, add one place where 'time' was missing and rename
the 'message' property to 'extra' property to match UnknownMessage's
declaration.

Don't set the 'filledSize' property in a TradeFinalizedMessage since
it isn't in the GDAXDoneMessage, the code was using a GDAXMatchMessage
for a done message.