GDAX factories check for required env vars before building GDAXAuthConfig.
This is a partial workaround to a problem where GDAXExchangeAPI will
construct a AuthenticatedClient if the constructor is given a defined
GDAXAuthConfig, however the constructor doesn't check that the key,
secret and passphrase properties are defined, so AuthenticatedClient
will throw an error.

So only a construct a GDAXAuthConfig if the required environmental
variables are all set.