GDAXMessage: switch to discriminated unions types.
Add a new GDAXMessage that is a union type and rename the old
GDAXMessage to OptionalUserId, dropping the 'GDAX' and the 'Message'
because the 'type property was removed since all sub-interfaces define
it exactly and the since the OptionalUserId super-interface isn't
meant to be publically used now.

Add GDAXHeartbeatMessage and GDAXReceivedMessage which is required to
avoid compiler errors with the current code.