GDAXFeed: remove many type assertions.
http://www.typescriptlang.org/docs/handbook/advanced-types.html

This is made possible by TypeScript's discriminated unions. For
example, handleMessage() calls isGDAXMessage(feedMessage) which means
that tsc knows that feedMessage is a GDAXMessage after the if
block. So switching on its type property means that the compiler knows
exactly the sub-interface in the case clause. So no downcasts needd.

Now that GDAXMessage is a union type, in a switch statement, the
TypeScript compiler knows the GDAXMessage specific type so no type
assertions are needed.