BaseError: new superclass to all Error classes.
I chose the name BaseError because GTTError was already taken, as
GTTError would be a natural name for the base class for all gdax-tt
Error classes. The name was also chosen to be consistent with ccxt's
base Error class.

This will let client code catch an error and use a single
'... instanceof BaseError' instead of three instanceof tests.