Breaking: All error classes now have an error cause property (#222)
* APIError: cause property is now an Error and add meta property.

Having all the gdax-trading-toolkit specific Error subclasses have a
consistent type for the cause property is very useful. To replace the
use of the cause property with non-Error instances add an optional
meta property.

* HTTPError: add an Error cause property.

HTTPErrorMessage's cause property is now the meta property. This will
allow the cause to turn into an Error later, for exception chaining.

* BaseError: new superclass to all Error classes.

I chose the name BaseError because GTTError was already taken, as
GTTError would be a natural name for the base class for all gdax-tt
Error classes. The name was also chosen to be consistent with ccxt's
base Error class.

This will let client code catch an error and use a single
'... instanceof BaseError' instead of three instanceof tests.

* ErrorMessage: cause property is now an optional Error.

At this point all the cause properties are Error objects which allows
for exception chaining.

* HTTPError: add unit tests for the clause property.